+++ source/extensions/lightspeed.event.shutdown_base/lightspeed/event/shutdown_base/__init__.py
@@ -15,4 +15,6 @@
 * limitations under the License.
 """
 
-from .extension import TrexStageCraftControlExtension
+__all__ = ["EventOnShutdownBase", "InterrupterBase"]
+
+from .base import EventOnShutdownBase, InterrupterBase