+++ source/extensions/lightspeed.event.shutdown_base/lightspeed/event/shutdown_base/tests/__init__.py
@@ -15,4 +15,4 @@
 * limitations under the License.
 """
 
-from .extension import TrexStageCraftControlExtension
+from .e2e import TestEventOnShutdownBase